Current Build on my machine
This build has the AI's seeking to the object whenever they are hit, in
order to test the way in which they handle hiding.  The final values for
when they need to hide and how that behavior conflicts with other
behavior choices has yet to be decided.

Hide has been turned into a composite goal with subgoals in order to
handle the movement to the object more easily.  Currently the only
needed sub goal is SeekToPosition.

RavenSteering contains a small amount of code for testing the movement
of objects around the obstacles.  This can be removed if/when the
current issues have been resolved.